empresasbr-API/requirements.txt
fastapi==0.104.0
SQLAlchemy==2.0.22
uvicorn==0.23.2
PyMySQL==1.1.0
python-dotenv==1.0.0